requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
pytest==7.4.2
python-dateutil==2.8.2
email_validator==2.0.0
